refactor: move checkpoints_exist utility to store
